Make the checks type-stable, and only checkbounds once - not for each component.

Speedup example:
```julia
julia> sa = StructArray(a=rand(1:10, 10^6), b=rand(10^6), c=["abc" for _ in 1:10^6], d=[:abc for _ in 1:10^6])

# before
julia> @btime view($sa, $([5, 10, 100]))
  195.191 ns (9 allocations: 624 bytes)

# after first commit
julia> @btime view($sa, $([5, 10, 100]))
  32.490 ns (0 allocations: 0 bytes)

# after second commit
julia> @btime view($sa, $([5, 10, 100]))
  20.309 ns (0 allocations: 0 bytes)

# still significantly slower than plain Vector, but much closer than before:
julia> @btime view($(collect(sa)), $([5, 10, 100]))
  7.353 ns (0 allocations: 0 bytes)
```