@piever 2022-05-01
Thanks, this is a nice change, I've left a couple of small comments.

As a general comment (though it should derail this PR, which is a nice performance improvement) one thing to consider is whether the `view`. overload is problematic. Maybe one should just have `view(s::StructArray, I...)` return a `SubArray{T, N, StructArray}` object. The order of double wrappers keeps confusing me (as in https://github.com/JuliaArrays/StructArrays.jl/pull/218#issue-1173957855) and is not entirely consistent at the moment.

Overall, it feels as though  wrapping each component is generally preferable, as it makes it easier to access the component vectors. Unfortunately, it is also the trickier to implement in general (esp. for `similar` and `reshape` with arbitrary ranges).

Do you have a strong opinion either way on this? Would one way work better than the other for your use case?