@aplavin 2022-05-01
> Do you have a strong opinion either way on this? Would one way work better than the other for your use case?

I strongly prefer the current situation, when each component is `view`ed, and they are wrapped into a top-level `StructArray`. As you say, it keeps the component access, which is quite an important `StructArrays` feature.

Would also be best for `StructArray + OffsetArray` to have the same order of wrappers, but that doesn't seem easy according to your comments in #218. Again, the motivation is that such an order gives both - offset indexing and component access - while the opposite order loses the latter.